Network.AWS.Discovery.CreateTags

Description

Creates one or more tags for configuration items. Tags are metadata that help you categorize IT assets. This API accepts a list of multiple configuration items.

Creating a Request

createTags :: CreateTags

Creates a value of CreateTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctConfigurationIds - A list of configuration items that you want to tag.

ctTags - Tags that you want to associate with one or more configuration items. Specify the tags that you want to create in a key -value format. For example: {"key": "serverType", "value": "webServer"}

data CreateTags

See: createTags smart constructor.

Request Lenses

ctConfigurationIds :: Lens' CreateTags [Text]

A list of configuration items that you want to tag.

ctTags :: Lens' CreateTags [Tag]

Tags that you want to associate with one or more configuration items. Specify the tags that you want to create in a key -value format. For example: {"key": "serverType", "value": "webServer"}

Destructuring the Response

createTagsResponse

Arguments

:: Int
ctrsResponseStatus

-> CreateTagsResponse

Creates a value of CreateTagsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctrsResponseStatus - -- | The response status code.

data CreateTagsResponse

See: createTagsResponse smart constructor.

Response Lenses

ctrsResponseStatus :: Lens' CreateTagsResponse Int

- | The response status code.