Network.AWS.Discovery.DescribeAgents

Description

Lists agents or the Connector by ID or lists all agents/Connectors associated with your user account if you did not specify an ID.

Creating a Request

describeAgents :: DescribeAgents

Creates a value of DescribeAgents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daAgentIds - The agent or the Connector IDs for which you want information. If you specify no IDs, the system returns information about all agents/Connectors associated with your AWS user account.

daFilters - You can filter the request using various logical operators and a key -value format. For example: {"key": "collectionStatus", "value": STARTED}

daNextToken - Token to retrieve the next set of results. For example, if you previously specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

daMaxResults - The total number of agents/Connectors to return in a single page of output. The maximum value is 100.

data DescribeAgents

See: describeAgents smart constructor.

Request Lenses

daAgentIds :: Lens' DescribeAgents [Text]

The agent or the Connector IDs for which you want information. If you specify no IDs, the system returns information about all agents/Connectors associated with your AWS user account.

daFilters :: Lens' DescribeAgents [Filter]

You can filter the request using various logical operators and a key -value format. For example: {"key": "collectionStatus", "value": STARTED}

daNextToken :: Lens' DescribeAgents (Maybe Text)

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

daMaxResults :: Lens' DescribeAgents (Maybe Int)

The total number of agents/Connectors to return in a single page of output. The maximum value is 100.

Destructuring the Response

describeAgentsResponse

Arguments

:: Int
dasrsResponseStatus

-> DescribeAgentsResponse

Creates a value of DescribeAgentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dasrsAgentsInfo - Lists agents or the Connector by ID or lists all agentsConnectors associated with your user account if you did not specify an agentConnector ID. The output includes agentConnector IDs, IP addresses, media access control (MAC) addresses, agentConnector health, host name where the agentConnector resides, and the version number of each agentConnector.

dasrsNextToken - Token to retrieve the next set of results. For example, if you specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

dasrsResponseStatus - -- | The response status code.

data DescribeAgentsResponse

See: describeAgentsResponse smart constructor.

Response Lenses

dasrsAgentsInfo :: Lens' DescribeAgentsResponse [AgentInfo]

Lists agents or the Connector by ID or lists all agentsConnectors associated with your user account if you did not specify an agentConnector ID. The output includes agentConnector IDs, IP addresses, media access control (MAC) addresses, agentConnector health, host name where the agentConnector resides, and the version number of each agentConnector.

dasrsNextToken :: Lens' DescribeAgentsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if you specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

dasrsResponseStatus :: Lens' DescribeAgentsResponse Int

- | The response status code.