Network.AWS.Discovery.DescribeExportTasks

Description

Retrieve status of one or more export tasks. You can retrieve the status of up to 100 export tasks.

Creating a Request

describeExportTasks :: DescribeExportTasks

Creates a value of DescribeExportTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detFilters - One or more filters. * AgentId - ID of the agent whose collected data will be exported

detNextToken - The nextToken value returned from a previous paginated DescribeExportTasks request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

detExportIds - One or more unique identifiers used to query the status of an export request.

detMaxResults - The maximum number of volume results returned by DescribeExportTasks in paginated output. When this parameter is used, DescribeExportTasks only returns maxResults results in a single page along with a nextToken response element.

data DescribeExportTasks

See: describeExportTasks smart constructor.

Request Lenses

detFilters :: Lens' DescribeExportTasks [ExportFilter]

One or more filters. * AgentId - ID of the agent whose collected data will be exported

detNextToken :: Lens' DescribeExportTasks (Maybe Text)

The nextToken value returned from a previous paginated DescribeExportTasks request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

detExportIds :: Lens' DescribeExportTasks [Text]

One or more unique identifiers used to query the status of an export request.

detMaxResults :: Lens' DescribeExportTasks (Maybe Int)

The maximum number of volume results returned by DescribeExportTasks in paginated output. When this parameter is used, DescribeExportTasks only returns maxResults results in a single page along with a nextToken response element.

Destructuring the Response

describeExportTasksResponse

Arguments

:: Int
detrsResponseStatus

-> DescribeExportTasksResponse

Creates a value of DescribeExportTasksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detrsNextToken - The nextToken value to include in a future DescribeExportTasks request. When the results of a DescribeExportTasks request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

detrsExportsInfo - Contains one or more sets of export request details. When the status of a request is SUCCEEDED , the response includes a URL for an Amazon S3 bucket where you can view the data in a CSV file.

detrsResponseStatus - -- | The response status code.

data DescribeExportTasksResponse

See: describeExportTasksResponse smart constructor.

Response Lenses

detrsNextToken :: Lens' DescribeExportTasksResponse (Maybe Text)

The nextToken value to include in a future DescribeExportTasks request. When the results of a DescribeExportTasks request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

detrsExportsInfo :: Lens' DescribeExportTasksResponse [ExportInfo]

Contains one or more sets of export request details. When the status of a request is SUCCEEDED , the response includes a URL for an Amazon S3 bucket where you can view the data in a CSV file.

detrsResponseStatus :: Lens' DescribeExportTasksResponse Int

- | The response status code.