Network.AWS.Discovery.GetDiscoverySummary

Description

Retrieves a short summary of discovered assets.

Creating a Request

getDiscoverySummary :: GetDiscoverySummary

Creates a value of GetDiscoverySummary with the minimum fields required to make a request.

data GetDiscoverySummary

See: getDiscoverySummary smart constructor.

Destructuring the Response

getDiscoverySummaryResponse

Arguments

:: Int
gdsrsResponseStatus

-> GetDiscoverySummaryResponse

Creates a value of GetDiscoverySummaryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsrsServers - The number of servers discovered.

gdsrsServersMappedtoTags - The number of servers mapped to tags.

gdsrsServersMappedToApplications - The number of servers mapped to applications.

gdsrsConnectorSummary - Details about discovered connectors, including connector status and health.

gdsrsAgentSummary - Details about discovered agents, including agent status and health.

gdsrsApplications - The number of applications discovered.

gdsrsResponseStatus - -- | The response status code.

data GetDiscoverySummaryResponse

See: getDiscoverySummaryResponse smart constructor.

Response Lenses

gdsrsServers :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers discovered.

gdsrsServersMappedtoTags :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers mapped to tags.

gdsrsServersMappedToApplications :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of servers mapped to applications.

gdsrsConnectorSummary :: Lens' GetDiscoverySummaryResponse (Maybe CustomerConnectorInfo)

Details about discovered connectors, including connector status and health.

gdsrsAgentSummary :: Lens' GetDiscoverySummaryResponse (Maybe CustomerAgentInfo)

Details about discovered agents, including agent status and health.

gdsrsApplications :: Lens' GetDiscoverySummaryResponse (Maybe Integer)

The number of applications discovered.

gdsrsResponseStatus :: Lens' GetDiscoverySummaryResponse Int

- | The response status code.