Network.AWS.Discovery.ListConfigurations

Description

Retrieves a list of configuration items according to criteria that you specify in a filter. The filter criteria identifies the relationship requirements.

Creating a Request

listConfigurations

Arguments

:: ConfigurationItemType
lcConfigurationType

-> ListConfigurations

Creates a value of ListConfigurations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcOrderBy - Certain filter criteria return output that can be sorted in ascending or descending order. For a list of output characteristics for each filter, see Using the ListConfigurations Action .

lcFilters - You can filter the request using various logical operators and a key -value format. For example: {"key": "serverType", "value": "webServer"} For a complete list of filter options and guidance about using them with this action, see Querying Discovered Configuration Items .

lcNextToken - Token to retrieve the next set of results. For example, if a previous call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

lcMaxResults - The total number of items to return. The maximum value is 100.

lcConfigurationType - A valid configuration identified by Application Discovery Service.

data ListConfigurations

See: listConfigurations smart constructor.

Request Lenses

lcOrderBy :: Lens' ListConfigurations [OrderByElement]

Certain filter criteria return output that can be sorted in ascending or descending order. For a list of output characteristics for each filter, see Using the ListConfigurations Action .

lcFilters :: Lens' ListConfigurations [Filter]

You can filter the request using various logical operators and a key -value format. For example: {"key": "serverType", "value": "webServer"} For a complete list of filter options and guidance about using them with this action, see Querying Discovered Configuration Items .

lcNextToken :: Lens' ListConfigurations (Maybe Text)

Token to retrieve the next set of results. For example, if a previous call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

lcMaxResults :: Lens' ListConfigurations (Maybe Int)

The total number of items to return. The maximum value is 100.

lcConfigurationType :: Lens' ListConfigurations ConfigurationItemType

A valid configuration identified by Application Discovery Service.

Destructuring the Response

listConfigurationsResponse

Arguments

:: Int
lcrsResponseStatus

-> ListConfigurationsResponse

Creates a value of ListConfigurationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcrsConfigurations - Returns configuration details, including the configuration ID, attribute names, and attribute values.

lcrsNextToken - Token to retrieve the next set of results. For example, if your call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

lcrsResponseStatus - -- | The response status code.

data ListConfigurationsResponse

See: listConfigurationsResponse smart constructor.

Response Lenses

lcrsConfigurations :: Lens' ListConfigurationsResponse [HashMap Text Text]

Returns configuration details, including the configuration ID, attribute names, and attribute values.

lcrsNextToken :: Lens' ListConfigurationsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if your call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

lcrsResponseStatus :: Lens' ListConfigurationsResponse Int

- | The response status code.