Network.AWS.Discovery.ListServerNeighbors

Description

Retrieves a list of servers that are one network hop away from a specified server.

Creating a Request

listServerNeighbors

Arguments

:: Text
lsnConfigurationId

-> ListServerNeighbors

Creates a value of ListServerNeighbors with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsnPortInformationNeeded - Flag to indicate if port and protocol information is needed as part of the response.

lsnNeighborConfigurationIds - List of configuration IDs to test for one-hop-away.

lsnNextToken - Token to retrieve the next set of results. For example, if you previously specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

lsnMaxResults - Maximum number of results to return in a single page of output.

lsnConfigurationId - Configuration ID of the server for which neighbors are being listed.

data ListServerNeighbors

See: listServerNeighbors smart constructor.

Request Lenses

lsnPortInformationNeeded :: Lens' ListServerNeighbors (Maybe Bool)

Flag to indicate if port and protocol information is needed as part of the response.

lsnNeighborConfigurationIds :: Lens' ListServerNeighbors [Text]

List of configuration IDs to test for one-hop-away.

lsnNextToken :: Lens' ListServerNeighbors (Maybe Text)

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

lsnMaxResults :: Lens' ListServerNeighbors (Maybe Int)

Maximum number of results to return in a single page of output.

lsnConfigurationId :: Lens' ListServerNeighbors Text

Configuration ID of the server for which neighbors are being listed.

Destructuring the Response

listServerNeighborsResponse

Arguments

:: Int
lsnrsResponseStatus

-> ListServerNeighborsResponse

Creates a value of ListServerNeighborsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsnrsNextToken - Token to retrieve the next set of results. For example, if you specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

lsnrsKnownDependencyCount - Count of distinct servers that are one hop away from the given server.

lsnrsResponseStatus - -- | The response status code.

lsnrsNeighbors - List of distinct servers that are one hop away from the given server.

data ListServerNeighborsResponse

See: listServerNeighborsResponse smart constructor.

Response Lenses

lsnrsNextToken :: Lens' ListServerNeighborsResponse (Maybe Text)

Token to retrieve the next set of results. For example, if you specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with this token. Use this token in the next query to retrieve the next set of 10.

lsnrsKnownDependencyCount :: Lens' ListServerNeighborsResponse (Maybe Integer)

Count of distinct servers that are one hop away from the given server.

lsnrsResponseStatus :: Lens' ListServerNeighborsResponse Int

- | The response status code.

lsnrsNeighbors :: Lens' ListServerNeighborsResponse [NeighborConnectionDetail]

List of distinct servers that are one hop away from the given server.