Network.AWS.Discovery.StartExportTask

Description

Begins the export of discovered data to an S3 bucket.

If you specify agentId in a filter, the task exports up to 72 hours of detailed data collected by the identified Application Discovery Agent, including network, process, and performance details. A time range for exported agent data may be set by using startTime and endTime . Export of detailed agent data is limited to five concurrently running exports.

If you do not include an agentId filter, summary data is exported that includes both AWS Agentless Discovery Connector data and summary data from AWS Discovery Agents. Export of summary data is limited to two exports per day.

Creating a Request

startExportTask :: StartExportTask

Creates a value of StartExportTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

setExportDataFormat - The file format for the returned export data. Default value is CSV .

setStartTime - The start timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, data is exported starting from the first data collected by the agent.

setFilters - If a filter is present, it selects the single agentId of the Application Discovery Agent for which data is exported. The agentId can be found in the results of the DescribeAgents API or CLI. If no filter is present, startTime and endTime are ignored and exported data includes both Agentless Discovery Connector data and summary data from Application Discovery agents.

setEndTime - The end timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, exported data includes the most recent data collected by the agent.

data StartExportTask

See: startExportTask smart constructor.

Request Lenses

setExportDataFormat :: Lens' StartExportTask [ExportDataFormat]

The file format for the returned export data. Default value is CSV .

setStartTime :: Lens' StartExportTask (Maybe UTCTime)

The start timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, data is exported starting from the first data collected by the agent.

setFilters :: Lens' StartExportTask [ExportFilter]

If a filter is present, it selects the single agentId of the Application Discovery Agent for which data is exported. The agentId can be found in the results of the DescribeAgents API or CLI. If no filter is present, startTime and endTime are ignored and exported data includes both Agentless Discovery Connector data and summary data from Application Discovery agents.

setEndTime :: Lens' StartExportTask (Maybe UTCTime)

The end timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, exported data includes the most recent data collected by the agent.

Destructuring the Response

startExportTaskResponse

Arguments

:: Int
setrsResponseStatus

-> StartExportTaskResponse

Creates a value of StartExportTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

setrsExportId - A unique identifier used to query the status of an export request.

setrsResponseStatus - -- | The response status code.

data StartExportTaskResponse

See: startExportTaskResponse smart constructor.

Response Lenses

setrsExportId :: Lens' StartExportTaskResponse (Maybe Text)

A unique identifier used to query the status of an export request.

setrsResponseStatus :: Lens' StartExportTaskResponse Int

- | The response status code.