Network.AWS.Discovery.Types

Description

Service Configuration

discovery :: Service

API version 2015-11-01 of the Amazon Application Discovery Service SDK configuration.

Errors

_AuthorizationErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS user account does not have permission to perform the action. Check the IAM policy associated with this account.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters are not valid. Verify the parameters and try again.

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

The value of one or more parameters are either invalid or out of range. Verify the parameter values and try again.

_ServerInternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The server experienced an internal error. Try again.

_OperationNotPermittedException :: AsError a => Getting (First ServiceError) a ServiceError

This operation is not permitted.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified configuration ID was not located. Verify the configuration ID and try again.

AgentStatus

data AgentStatus

Constructors

Blacklisted

Healthy

Running

Shutdown

Unhealthy

Unknown

ConfigurationItemType

data ConfigurationItemType

Constructors

Application

Connection

Process

Server

ExportDataFormat

data ExportDataFormat

Constructors

CSV

Graphml

ExportStatus

data ExportStatus

Constructors

Failed

InProgress

Succeeded

OrderString

data OrderString

Constructors

Asc

Desc

AgentConfigurationStatus

data AgentConfigurationStatus

Information about agents or connectors that were instructed to start collecting data. Information includes the agentconnector ID, a description of the operation, and whether the agentconnector configuration was updated.

See: agentConfigurationStatus smart constructor.

agentConfigurationStatus :: AgentConfigurationStatus

Creates a value of AgentConfigurationStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acsAgentId - The agent/connector ID.

acsOperationSucceeded - Information about the status of the StartDataCollection and StopDataCollection operations. The system has recorded the data collection operation. The agent/connector receives this command the next time it polls for a new command.

acsDescription - A description of the operation performed.

acsAgentId :: Lens' AgentConfigurationStatus (Maybe Text)

The agent/connector ID.

acsOperationSucceeded :: Lens' AgentConfigurationStatus (Maybe Bool)

Information about the status of the StartDataCollection and StopDataCollection operations. The system has recorded the data collection operation. The agent/connector receives this command the next time it polls for a new command.

acsDescription :: Lens' AgentConfigurationStatus (Maybe Text)

A description of the operation performed.

AgentInfo

data AgentInfo

Information about agents or connectors associated with the user’s AWS account. Information includes agent or connector IDs, IP addresses, media access control (MAC) addresses, agent or connector health, hostname where the agent or connector resides, and agent version for each agent.

See: agentInfo smart constructor.

agentInfo :: AgentInfo

Creates a value of AgentInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aiHostName - The name of the host where the agent or connector resides. The host can be a server or virtual machine.

aiLastHealthPingTime - Time since agent or connector health was reported.

aiAgentNetworkInfoList - Network details about the host where the agent or connector resides.

aiConnectorId - The ID of the connector.

aiHealth - The health of the agent or connector.

aiAgentId - The agent or connector ID.

aiVersion - The agent or connector version.

aiCollectionStatus - Status of the collection process for an agent or connector.

aiRegisteredTime - Agent's first registration timestamp in UTC.

aiAgentType - Type of agent.

aiHostName :: Lens' AgentInfo (Maybe Text)

The name of the host where the agent or connector resides. The host can be a server or virtual machine.

aiLastHealthPingTime :: Lens' AgentInfo (Maybe Text)

Time since agent or connector health was reported.

aiAgentNetworkInfoList :: Lens' AgentInfo [AgentNetworkInfo]

Network details about the host where the agent or connector resides.

aiConnectorId :: Lens' AgentInfo (Maybe Text)

The ID of the connector.

aiHealth :: Lens' AgentInfo (Maybe AgentStatus)

The health of the agent or connector.

aiAgentId :: Lens' AgentInfo (Maybe Text)

The agent or connector ID.

aiVersion :: Lens' AgentInfo (Maybe Text)

The agent or connector version.

aiCollectionStatus :: Lens' AgentInfo (Maybe Text)

Status of the collection process for an agent or connector.

aiRegisteredTime :: Lens' AgentInfo (Maybe Text)

Agent's first registration timestamp in UTC.

aiAgentType :: Lens' AgentInfo (Maybe Text)

Type of agent.

AgentNetworkInfo

data AgentNetworkInfo

Network details about the host where the agent/connector resides.

See: agentNetworkInfo smart constructor.

agentNetworkInfo :: AgentNetworkInfo

Creates a value of AgentNetworkInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aniIpAddress - The IP address for the host where the agent/connector resides.

aniMacAddress - The MAC address for the host where the agent/connector resides.

aniIpAddress :: Lens' AgentNetworkInfo (Maybe Text)

The IP address for the host where the agent/connector resides.

aniMacAddress :: Lens' AgentNetworkInfo (Maybe Text)

The MAC address for the host where the agent/connector resides.

ConfigurationTag

data ConfigurationTag

Tags for a configuration item. Tags are metadata that help you categorize IT assets.

See: configurationTag smart constructor.

configurationTag :: ConfigurationTag

Creates a value of ConfigurationTag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctTimeOfCreation - The time the configuration tag was created in Coordinated Universal Time (UTC).

ctConfigurationId - The configuration ID for the item to tag. You can specify a list of keys and values.

ctConfigurationType - A type of IT asset to tag.

ctValue - A value on which to filter. For example key = serverType and value = web server .

ctKey - A type of tag on which to filter. For example, serverType .

ctTimeOfCreation :: Lens' ConfigurationTag (Maybe UTCTime)

The time the configuration tag was created in Coordinated Universal Time (UTC).

ctConfigurationId :: Lens' ConfigurationTag (Maybe Text)

The configuration ID for the item to tag. You can specify a list of keys and values.

ctConfigurationType :: Lens' ConfigurationTag (Maybe ConfigurationItemType)

A type of IT asset to tag.

ctValue :: Lens' ConfigurationTag (Maybe Text)

A value on which to filter. For example key = serverType and value = web server .

ctKey :: Lens' ConfigurationTag (Maybe Text)

A type of tag on which to filter. For example, serverType .

CustomerAgentInfo

data CustomerAgentInfo

Inventory data for installed discovery agents.

See: customerAgentInfo smart constructor.

customerAgentInfo

Arguments

:: Int
caiActiveAgents

-> Int
caiHealthyAgents

-> Int
caiBlackListedAgents

-> Int
caiShutdownAgents

-> Int
caiUnhealthyAgents

-> Int
caiTotalAgents

-> Int
caiUnknownAgents

-> CustomerAgentInfo

Creates a value of CustomerAgentInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caiActiveAgents - Number of active discovery agents.

caiHealthyAgents - Number of healthy discovery agents

caiBlackListedAgents - Number of blacklisted discovery agents.

caiShutdownAgents - Number of discovery agents with status SHUTDOWN.

caiUnhealthyAgents - Number of unhealthy discovery agents.

caiTotalAgents - Total number of discovery agents.

caiUnknownAgents - Number of unknown discovery agents.

caiActiveAgents :: Lens' CustomerAgentInfo Int

Number of active discovery agents.

caiHealthyAgents :: Lens' CustomerAgentInfo Int

Number of healthy discovery agents

caiBlackListedAgents :: Lens' CustomerAgentInfo Int

Number of blacklisted discovery agents.

caiShutdownAgents :: Lens' CustomerAgentInfo Int

Number of discovery agents with status SHUTDOWN.

caiUnhealthyAgents :: Lens' CustomerAgentInfo Int

Number of unhealthy discovery agents.

caiTotalAgents :: Lens' CustomerAgentInfo Int

Total number of discovery agents.

caiUnknownAgents :: Lens' CustomerAgentInfo Int

Number of unknown discovery agents.

CustomerConnectorInfo

data CustomerConnectorInfo

Inventory data for installed discovery connectors.

See: customerConnectorInfo smart constructor.

customerConnectorInfo

Arguments

:: Int
cciActiveConnectors

-> Int
cciHealthyConnectors

-> Int
cciBlackListedConnectors

-> Int
cciShutdownConnectors

-> Int
cciUnhealthyConnectors

-> Int
cciTotalConnectors

-> Int
cciUnknownConnectors

-> CustomerConnectorInfo

Creates a value of CustomerConnectorInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cciActiveConnectors - Number of active discovery connectors.

cciHealthyConnectors - Number of healthy discovery connectors.

cciBlackListedConnectors - Number of blacklisted discovery connectors.

cciShutdownConnectors - Number of discovery connectors with status SHUTDOWN,

cciUnhealthyConnectors - Number of unhealthy discovery connectors.

cciTotalConnectors - Total number of discovery connectors.

cciUnknownConnectors - Number of unknown discovery connectors.

cciActiveConnectors :: Lens' CustomerConnectorInfo Int

Number of active discovery connectors.

cciHealthyConnectors :: Lens' CustomerConnectorInfo Int

Number of healthy discovery connectors.

cciBlackListedConnectors :: Lens' CustomerConnectorInfo Int

Number of blacklisted discovery connectors.

cciShutdownConnectors :: Lens' CustomerConnectorInfo Int

Number of discovery connectors with status SHUTDOWN,

cciUnhealthyConnectors :: Lens' CustomerConnectorInfo Int

Number of unhealthy discovery connectors.

cciTotalConnectors :: Lens' CustomerConnectorInfo Int

Total number of discovery connectors.

cciUnknownConnectors :: Lens' CustomerConnectorInfo Int

Number of unknown discovery connectors.

ExportFilter

data ExportFilter

Used to select which agent's data is to be exported. A single agent ID may be selected for export using the StartExportTask action.

See: exportFilter smart constructor.

exportFilter

Arguments

:: Text
efName

-> Text
efCondition

-> ExportFilter

Creates a value of ExportFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

efName - A single ExportFilter name. Supported filters: agentId .

efValues - A single agentId for a Discovery Agent. An agentId can be found using the DescribeAgents action. Typically an ADS agentId is in the form o-0123456789abcdef0 .

efCondition - Supported condition: EQUALS

efName :: Lens' ExportFilter Text

A single ExportFilter name. Supported filters: agentId .

efValues :: Lens' ExportFilter [Text]

A single agentId for a Discovery Agent. An agentId can be found using the DescribeAgents action. Typically an ADS agentId is in the form o-0123456789abcdef0 .

efCondition :: Lens' ExportFilter Text

Supported condition: EQUALS

ExportInfo

data ExportInfo

Information regarding the export status of discovered data. The value is an array of objects.

See: exportInfo smart constructor.

exportInfo

Arguments

:: Text
eiExportId

-> ExportStatus
eiExportStatus

-> Text
eiStatusMessage

-> UTCTime
eiExportRequestTime

-> ExportInfo

Creates a value of ExportInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eiConfigurationsDownloadURL - A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

eiRequestedStartTime - The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo .

eiRequestedEndTime - The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo .

eiIsTruncated - If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate .

eiExportId - A unique identifier used to query an export.

eiExportStatus - The status of the data export job.

eiStatusMessage - A status message provided for API callers.

eiExportRequestTime - The time that the data export was initiated.

eiConfigurationsDownloadURL :: Lens' ExportInfo (Maybe Text)

A URL for an Amazon S3 bucket where you can review the exported data. The URL is displayed only if the export succeeded.

eiRequestedStartTime :: Lens' ExportInfo (Maybe UTCTime)

The value of startTime parameter in the StartExportTask request. If no startTime was requested, this result does not appear in ExportInfo .

eiRequestedEndTime :: Lens' ExportInfo (Maybe UTCTime)

The endTime used in the StartExportTask request. If no endTime was requested, this result does not appear in ExportInfo .

eiIsTruncated :: Lens' ExportInfo (Maybe Bool)

If true, the export of agent information exceeded the size limit for a single export and the exported data is incomplete for the requested time range. To address this, select a smaller time range for the export by using startDate and endDate .

eiExportId :: Lens' ExportInfo Text

A unique identifier used to query an export.

eiExportStatus :: Lens' ExportInfo ExportStatus

The status of the data export job.

eiStatusMessage :: Lens' ExportInfo Text

A status message provided for API callers.

eiExportRequestTime :: Lens' ExportInfo UTCTime

The time that the data export was initiated.

Filter

data Filter

A filter that can use conditional operators.

For more information about filters, see Querying Discovered Configuration Items .

See: filter' smart constructor.

filter'

Arguments

:: Text
fName

-> Text
fCondition

-> Filter

Creates a value of Filter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fName - The name of the filter.

fValues - A string value on which to filter. For example, if you choose the destinationServer.osVersion filter name, you could specify Ubuntu for the value.

fCondition - A conditional operator. The following operators are valid: EQUALS, NOT_EQUALS, CONTAINS, NOT_CONTAINS. If you specify multiple filters, the system utilizes all filters as though concatenated by AND . If you specify multiple values for a particular filter, the system differentiates the values using OR . Calling either DescribeConfigurations or ListConfigurations returns attributes of matching configuration items.

fName :: Lens' Filter Text

The name of the filter.

fValues :: Lens' Filter [Text]

A string value on which to filter. For example, if you choose the destinationServer.osVersion filter name, you could specify Ubuntu for the value.

fCondition :: Lens' Filter Text

A conditional operator. The following operators are valid: EQUALS, NOT_EQUALS, CONTAINS, NOT_CONTAINS. If you specify multiple filters, the system utilizes all filters as though concatenated by AND . If you specify multiple values for a particular filter, the system differentiates the values using OR . Calling either DescribeConfigurations or ListConfigurations returns attributes of matching configuration items.

NeighborConnectionDetail

data NeighborConnectionDetail

Details about neighboring servers.

See: neighborConnectionDetail smart constructor.

neighborConnectionDetail

Arguments

:: Text
ncdSourceServerId

-> Text
ncdDestinationServerId

-> Integer
ncdConnectionsCount

-> NeighborConnectionDetail

Creates a value of NeighborConnectionDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ncdTransportProtocol - The network protocol used for the connection.

ncdDestinationPort - The destination network port for the connection.

ncdSourceServerId - The ID of the server that opened the network connection.

ncdDestinationServerId - The ID of the server that accepted the network connection.

ncdConnectionsCount - The number of open network connections with the neighboring server.

ncdTransportProtocol :: Lens' NeighborConnectionDetail (Maybe Text)

The network protocol used for the connection.

ncdDestinationPort :: Lens' NeighborConnectionDetail (Maybe Int)

The destination network port for the connection.

ncdSourceServerId :: Lens' NeighborConnectionDetail Text

The ID of the server that opened the network connection.

ncdDestinationServerId :: Lens' NeighborConnectionDetail Text

The ID of the server that accepted the network connection.

ncdConnectionsCount :: Lens' NeighborConnectionDetail Integer

The number of open network connections with the neighboring server.

OrderByElement

data OrderByElement

A field and direction for ordered output.

See: orderByElement smart constructor.

orderByElement

Arguments

:: Text
obeFieldName

-> OrderByElement

Creates a value of OrderByElement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

obeSortOrder - Ordering direction.

obeFieldName - The field on which to order.

obeSortOrder :: Lens' OrderByElement (Maybe OrderString)

Ordering direction.

obeFieldName :: Lens' OrderByElement Text

The field on which to order.

Tag

data Tag

Metadata that help you categorize IT assets.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - The type of tag on which to filter.

tagValue - A value for a tag key on which to filter.

tagKey :: Lens' Tag Text

The type of tag on which to filter.

tagValue :: Lens' Tag Text

A value for a tag key on which to filter.

TagFilter

data TagFilter

The tag filter. Valid names are: tagKey , tagValue , configurationId .

See: tagFilter smart constructor.

tagFilter

Arguments

:: Text
tfName

-> TagFilter

Creates a value of TagFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tfName - A name of the tag filter.

tfValues - Values for the tag filter.

tfName :: Lens' TagFilter Text

A name of the tag filter.

tfValues :: Lens' TagFilter [Text]

Values for the tag filter.